+++ rpc/tendermint/test/websocket_client_test.go
@@ -7,12 +7,13 @@ import (
 	"fmt"
 	"testing"
 
+	"time"
+
+	core_types "github.com/eris-ltd/eris-db/rpc/tendermint/core/types"
 	"github.com/eris-ltd/eris-db/txs"
 	"github.com/stretchr/testify/assert"
 	_ "github.com/tendermint/tendermint/config/tendermint_test"
 )
-
-var wsTyp = "JSONRPC"
 
 //--------------------------------------------------------------------------------
 // Test the websocket service
@@ -91,8 +92,8 @@ func TestWSSend(t *testing.T) {
 		wsc.Stop()
 	}()
 	waitForEvent(t, wsc, eidInput, func() {
-		tx := makeDefaultSendTxSigned(t, wsTyp, toAddr, amt)
-		broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultSendTxSigned(t, jsonRpcClient, toAddr, amt)
+		broadcastTx(t, jsonRpcClient, tx)
 	}, unmarshalValidateSend(amt, toAddr))
 
 	waitForEvent(t, wsc, eidOutput, func() {},
@@ -115,8 +116,8 @@ func TestWSDoubleFire(t *testing.T) {
 	toAddr := user[1].Address
 	// broadcast the transaction, wait to hear about it
 	waitForEvent(t, wsc, eid, func() {
-		tx := makeDefaultSendTxSigned(t, wsTyp, toAddr, amt)
-		broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultSendTxSigned(t, jsonRpcClient, toAddr, amt)
+		broadcastTx(t, jsonRpcClient, tx)
 	}, func(eid string, b txs.EventData) (bool, error) {
 		return true, nil
 	})
@@ -147,8 +148,8 @@ func TestWSCallWait(t *testing.T) {
 	var contractAddr []byte
 	// wait for the contract to be created
 	waitForEvent(t, wsc, eid1, func() {
-		tx := makeDefaultCallTx(t, wsTyp, nil, code, amt, gasLim, fee)
-		receipt := broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultCallTx(t, jsonRpcClient, nil, code, amt, gasLim, fee)
+		receipt := broadcastTx(t, jsonRpcClient, tx)
 		contractAddr = receipt.ContractAddr
 	}, unmarshalValidateTx(amt, returnCode))
 
@@ -162,8 +163,8 @@ func TestWSCallWait(t *testing.T) {
 	// get the return value from a call
 	data := []byte{0x1}
 	waitForEvent(t, wsc, eid2, func() {
-		tx := makeDefaultCallTx(t, wsTyp, contractAddr, data, amt, gasLim, fee)
-		receipt := broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultCallTx(t, jsonRpcClient, contractAddr, data, amt, gasLim, fee)
+		receipt := broadcastTx(t, jsonRpcClient, tx)
 		contractAddr = receipt.ContractAddr
 	}, unmarshalValidateTx(amt, returnVal))
 }
@@ -178,8 +179,8 @@ func TestWSCallNoWait(t *testing.T) {
 	amt, gasLim, fee := int64(10000), int64(1000), int64(1000)
 	code, _, returnVal := simpleContract()
 
-	tx := makeDefaultCallTx(t, wsTyp, nil, code, amt, gasLim, fee)
-	receipt := broadcastTx(t, wsTyp, tx)
+	tx := makeDefaultCallTx(t, jsonRpcClient, nil, code, amt, gasLim, fee)
+	receipt := broadcastTx(t, jsonRpcClient, tx)
 	contractAddr := receipt.ContractAddr
 
 	// susbscribe to the new contract
@@ -193,8 +194,8 @@ func TestWSCallNoWait(t *testing.T) {
 	// get the return value from a call
 	data := []byte{0x1}
 	waitForEvent(t, wsc, eid, func() {
-		tx := makeDefaultCallTx(t, wsTyp, contractAddr, data, amt, gasLim, fee)
-		broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultCallTx(t, jsonRpcClient, contractAddr, data, amt, gasLim, fee)
+		broadcastTx(t, jsonRpcClient, tx)
 	}, unmarshalValidateTx(amt, returnVal))
 }
 
@@ -209,13 +210,13 @@ func TestWSCallCall(t *testing.T) {
 	txid := new([]byte)
 
 	// deploy the two contracts
-	tx := makeDefaultCallTx(t, wsTyp, nil, code, amt, gasLim, fee)
-	receipt := broadcastTx(t, wsTyp, tx)
+	tx := makeDefaultCallTx(t, jsonRpcClient, nil, code, amt, gasLim, fee)
+	receipt := broadcastTx(t, jsonRpcClient, tx)
 	contractAddr1 := receipt.ContractAddr
 
 	code, _, _ = simpleCallContract(contractAddr1)
-	tx = makeDefaultCallTx(t, wsTyp, nil, code, amt, gasLim, fee)
-	receipt = broadcastTx(t, wsTyp, tx)
+	tx = makeDefaultCallTx(t, jsonRpcClient, nil, code, amt, gasLim, fee)
+	receipt = broadcastTx(t, jsonRpcClient, tx)
 	contractAddr2 := receipt.ContractAddr
 
 	// subscribe to the new contracts
@@ -235,12 +236,59 @@ func TestWSCallCall(t *testing.T) {
 	})
 	// call it
 	waitForEvent(t, wsc, eid, func() {
-		tx := makeDefaultCallTx(t, wsTyp, contractAddr2, nil, amt, gasLim, fee)
-		broadcastTx(t, wsTyp, tx)
+		tx := makeDefaultCallTx(t, jsonRpcClient, contractAddr2, nil, amt, gasLim, fee)
+		broadcastTx(t, jsonRpcClient, tx)
 		*txid = txs.TxHash(chainID, tx)
 	}, unmarshalValidateCall(user[0].Address, returnVal, txid))
 }
 
 func TestSubscribe(t *testing.T) {
-	testSubscribe(t)
+	var subId string
+	wsc := newWSClient(t)
+	subscribe(t, wsc, txs.EventStringNewBlock())
+
+	timeout := time.NewTimer(timeoutSeconds * time.Second)
+Subscribe:
+	for {
+		select {
+		case <-timeout.C:
+			t.Fatal("Timed out waiting for subscription result")
+
+		case bs := <-wsc.ResultsCh:
+			resultSubscribe, ok := readResult(t, bs).(*core_types.ResultSubscribe)
+			if ok {
+				assert.Equal(t, txs.EventStringNewBlock(), resultSubscribe.Event)
+				subId = resultSubscribe.SubscriptionId
+				break Subscribe
+			}
+		}
+	}
+
+	seenBlock := false
+	timeout = time.NewTimer(timeoutSeconds * time.Second)
+	for {
+		select {
+		case <-timeout.C:
+			if !seenBlock {
+				t.Fatal("Timed out without seeing a NewBlock event")
+			}
+			return
+
+		case bs := <-wsc.ResultsCh:
+			resultEvent, ok := readResult(t, bs).(*core_types.ResultEvent)
+			if ok {
+				_, ok := resultEvent.Data.(txs.EventDataNewBlock)
+				if ok {
+					if seenBlock {
+						// There's a mild race here, but when we enter we've just seen a block
+						// so we should be able to unsubscribe before we see another block
+						t.Fatal("Continued to see NewBlock event after unsubscribing")
+					} else {
+						seenBlock = true
+						unsubscribe(t, wsc, subId)
+					}
+				}
+			}
+		}
+	}
 }
